Nest all MathML element types inside the Mathml object
Archimedes exported ~65 MathML element names flat into the `soundness`
package — including generic words like `Token`, `Layout`, `Tabular` and `Ms`
that collided with unrelated modules (harlequin's `Token`, punctuation's
`Layout`, escritoire's `Tabular`, cataclysm's millisecond `Ms`) under
`import soundness.*`.

All node types — the six category traits and every concrete element case
class with its companion — now live inside `object Mathml` (as `Mathml.Mi`,
`Mathml.Mrow`, `Mathml.Token`, …). Since a Scala 3 object cannot span files,
the six per-category element files are consolidated into
`archimedes.Mathml.scala`. Only the API surface (`Math`, `Mathml`, `Display`,
the parser/reader/errors, `Cell`, and the term helpers) is exported flat;
element-constructing sites gain `import Mathml.*`, and callers author bare
element names via `import soundness.Mathml.*`.

Behaviour is unchanged; all 80 archimedes tests pass.
