--- lib/archimedes/src/core/archimedes.Elementary.scala
@@ -1,86 +1,0 @@
-                                                                                                  /*
-┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┓
-┃                                                                                                  ┃
-┃                                                   ╭───╮                                          ┃
-┃                                                   │   │                                          ┃
-┃                                                   │   │                                          ┃
-┃   ╭───────╮╭─────────╮╭───╮ ╭───╮╭───╮╌────╮╭────╌┤   │╭───╮╌────╮╭────────╮╭───────╮╭───────╮   ┃
-┃   │   ╭───╯│   ╭─╮   ││   │ │   ││   ╭─╮   ││   ╭─╮   ││   ╭─╮   ││   ╭─╮  ││   ╭───╯│   ╭───╯   ┃
-┃   │   ╰───╮│   │ │   ││   │ │   ││   │ │   ││   │ │   ││   │ │   ││   ╰─╯  ││   ╰───╮│   ╰───╮   ┃
-┃   ╰───╮   ││   │ │   ││   │ │   ││   │ │   ││   │ │   ││   │ │   ││   ╭────╯╰───╮   │╰───╮   │   ┃
-┃   ╭───╯   ││   ╰─╯   ││   ╰─╯   ││   │ │   ││   ╰─╯   ││   │ │   ││   ╰────╮╭───╯   │╭───╯   │   ┃
-┃   ╰───────╯╰─────────╯╰────╌╰───╯╰───╯ ╰───╯╰────╌╰───╯╰───╯ ╰───╯╰────────╯╰───────╯╰───────╯   ┃
-┃                                                                                                  ┃
-┃    Soundness, version 0.63.0.                                                                    ┃
-┃    © Copyright 2021-25 Jon Pretty, Propensive OÜ.                                                ┃
-┃                                                                                                  ┃
-┃    The primary distribution site is:                                                             ┃
-┃                                                                                                  ┃
-┃        https://soundness.dev/                                                                    ┃
-┃                                                                                                  ┃
-┃    Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file     ┃
-┃    except in compliance with the License. You may obtain a copy of the License at                ┃
-┃                                                                                                  ┃
-┃        https://www.apache.org/licenses/LICENSE-2.0                                               ┃
-┃                                                                                                  ┃
-┃    Unless required by applicable law or agreed to in writing,  software distributed under the    ┃
-┃    License is distributed on an "AS IS" BASIS,  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,    ┃
-┃    either express or implied. See the License for the specific language governing permissions    ┃
-┃    and limitations under the License.                                                            ┃
-┃                                                                                                  ┃
-┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┛
-                                                                                                  */
-package archimedes
-
-import anticipation.*
-import gossamer.*
-import vacuous.*
-
-// Elementary-math schemata, used for column arithmetic and long division:
-// `<mstack>`, `<mlongdiv>`, `<msgroup>`, `<msrow>`, `<mscarries>`, `<mscarry>`
-// and `<msline>`.
-
-sealed trait Elementary extends Mathml:
-  def text: Optional[Text] = Unset
-
-object Mstack:
-  def apply(children: Mathml*): Mstack = Mstack(children.to(List))
-
-case class Mstack(contents: List[Mathml], attributes: List[(Text, Text)] = Nil) extends Elementary:
-  def label: Text = t"mstack"
-
-object Mlongdiv:
-  def apply(children: Mathml*): Mlongdiv = Mlongdiv(children.to(List))
-
-case class Mlongdiv(contents: List[Mathml], attributes: List[(Text, Text)] = Nil)
-extends Elementary:
-  def label: Text = t"mlongdiv"
-
-object Msgroup:
-  def apply(children: Mathml*): Msgroup = Msgroup(children.to(List))
-
-case class Msgroup(contents: List[Mathml], attributes: List[(Text, Text)] = Nil) extends Elementary:
-  def label: Text = t"msgroup"
-
-object Msrow:
-  def apply(children: Mathml*): Msrow = Msrow(children.to(List))
-
-case class Msrow(contents: List[Mathml], attributes: List[(Text, Text)] = Nil) extends Elementary:
-  def label: Text = t"msrow"
-
-object Mscarries:
-  def apply(children: Mathml*): Mscarries = Mscarries(children.to(List))
-
-case class Mscarries(contents: List[Mathml], attributes: List[(Text, Text)] = Nil)
-extends Elementary:
-  def label: Text = t"mscarries"
-
-object Mscarry:
-  def apply(children: Mathml*): Mscarry = Mscarry(children.to(List))
-
-case class Mscarry(contents: List[Mathml], attributes: List[(Text, Text)] = Nil) extends Elementary:
-  def label: Text = t"mscarry"
-
-case class Msline(attributes: List[(Text, Text)] = Nil) extends Elementary:
-  def label: Text = t"msline"
-  def contents: List[Mathml] = Nil
