--- lib/archimedes/src/core/archimedes.Semantic.scala
@@ -1,71 +1,0 @@
-                                                                                                  /*
-┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┓
-┃                                                                                                  ┃
-┃                                                   ╭───╮                                          ┃
-┃                                                   │   │                                          ┃
-┃                                                   │   │                                          ┃
-┃   ╭───────╮╭─────────╮╭───╮ ╭───╮╭───╮╌────╮╭────╌┤   │╭───╮╌────╮╭────────╮╭───────╮╭───────╮   ┃
-┃   │   ╭───╯│   ╭─╮   ││   │ │   ││   ╭─╮   ││   ╭─╮   ││   ╭─╮   ││   ╭─╮  ││   ╭───╯│   ╭───╯   ┃
-┃   │   ╰───╮│   │ │   ││   │ │   ││   │ │   ││   │ │   ││   │ │   ││   ╰─╯  ││   ╰───╮│   ╰───╮   ┃
-┃   ╰───╮   ││   │ │   ││   │ │   ││   │ │   ││   │ │   ││   │ │   ││   ╭────╯╰───╮   │╰───╮   │   ┃
-┃   ╭───╯   ││   ╰─╯   ││   ╰─╯   ││   │ │   ││   ╰─╯   ││   │ │   ││   ╰────╮╭───╯   │╭───╯   │   ┃
-┃   ╰───────╯╰─────────╯╰────╌╰───╯╰───╯ ╰───╯╰────╌╰───╯╰───╯ ╰───╯╰────────╯╰───────╯╰───────╯   ┃
-┃                                                                                                  ┃
-┃    Soundness, version 0.63.0.                                                                    ┃
-┃    © Copyright 2021-25 Jon Pretty, Propensive OÜ.                                                ┃
-┃                                                                                                  ┃
-┃    The primary distribution site is:                                                             ┃
-┃                                                                                                  ┃
-┃        https://soundness.dev/                                                                    ┃
-┃                                                                                                  ┃
-┃    Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file     ┃
-┃    except in compliance with the License. You may obtain a copy of the License at                ┃
-┃                                                                                                  ┃
-┃        https://www.apache.org/licenses/LICENSE-2.0                                               ┃
-┃                                                                                                  ┃
-┃    Unless required by applicable law or agreed to in writing,  software distributed under the    ┃
-┃    License is distributed on an "AS IS" BASIS,  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,    ┃
-┃    either express or implied. See the License for the specific language governing permissions    ┃
-┃    and limitations under the License.                                                            ┃
-┃                                                                                                  ┃
-┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┛
-                                                                                                  */
-package archimedes
-
-import anticipation.*
-import gossamer.*
-import vacuous.*
-
-// `<maction>` (bound actions such as toggle/highlight) plus the semantics
-// bridge: `<semantics>` pairs a presentation subtree with one or more
-// annotations. `<annotation>` carries character data (e.g. a TeX string) while
-// `<annotation-xml>` carries a nested markup subtree.
-
-sealed trait Semantic extends Mathml
-
-object Maction:
-  def apply(children: Mathml*): Maction = Maction(children.to(List))
-
-case class Maction(contents: List[Mathml], attributes: List[(Text, Text)] = Nil) extends Semantic:
-  def label: Text = t"maction"
-  def text: Optional[Text] = Unset
-
-object Semantics:
-  def apply(children: Mathml*): Semantics = Semantics(children.to(List))
-
-case class Semantics(contents: List[Mathml], attributes: List[(Text, Text)] = Nil) extends Semantic:
-  def label: Text = t"semantics"
-  def text: Optional[Text] = Unset
-
-case class Annotation(value: Text, attributes: List[(Text, Text)] = Nil) extends Semantic:
-  def label: Text = t"annotation"
-  def contents: List[Mathml] = Nil
-  def text: Optional[Text] = value
-
-object AnnotationXml:
-  def apply(children: Mathml*): AnnotationXml = AnnotationXml(children.to(List))
-
-case class AnnotationXml(contents: List[Mathml], attributes: List[(Text, Text)] = Nil)
-extends Semantic:
-  def label: Text = t"annotation-xml"
-  def text: Optional[Text] = Unset
