--- lib/archimedes/src/core/archimedes.Token.scala
@@ -1,72 +1,0 @@
-                                                                                                  /*
-┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┓
-┃                                                                                                  ┃
-┃                                                   ╭───╮                                          ┃
-┃                                                   │   │                                          ┃
-┃                                                   │   │                                          ┃
-┃   ╭───────╮╭─────────╮╭───╮ ╭───╮╭───╮╌────╮╭────╌┤   │╭───╮╌────╮╭────────╮╭───────╮╭───────╮   ┃
-┃   │   ╭───╯│   ╭─╮   ││   │ │   ││   ╭─╮   ││   ╭─╮   ││   ╭─╮   ││   ╭─╮  ││   ╭───╯│   ╭───╯   ┃
-┃   │   ╰───╮│   │ │   ││   │ │   ││   │ │   ││   │ │   ││   │ │   ││   ╰─╯  ││   ╰───╮│   ╰───╮   ┃
-┃   ╰───╮   ││   │ │   ││   │ │   ││   │ │   ││   │ │   ││   │ │   ││   ╭────╯╰───╮   │╰───╮   │   ┃
-┃   ╭───╯   ││   ╰─╯   ││   ╰─╯   ││   │ │   ││   ╰─╯   ││   │ │   ││   ╰────╮╭───╯   │╭───╯   │   ┃
-┃   ╰───────╯╰─────────╯╰────╌╰───╯╰───╯ ╰───╯╰────╌╰───╯╰───╯ ╰───╯╰────────╯╰───────╯╰───────╯   ┃
-┃                                                                                                  ┃
-┃    Soundness, version 0.63.0.                                                                    ┃
-┃    © Copyright 2021-25 Jon Pretty, Propensive OÜ.                                                ┃
-┃                                                                                                  ┃
-┃    The primary distribution site is:                                                             ┃
-┃                                                                                                  ┃
-┃        https://soundness.dev/                                                                    ┃
-┃                                                                                                  ┃
-┃    Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file     ┃
-┃    except in compliance with the License. You may obtain a copy of the License at                ┃
-┃                                                                                                  ┃
-┃        https://www.apache.org/licenses/LICENSE-2.0                                               ┃
-┃                                                                                                  ┃
-┃    Unless required by applicable law or agreed to in writing,  software distributed under the    ┃
-┃    License is distributed on an "AS IS" BASIS,  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,    ┃
-┃    either express or implied. See the License for the specific language governing permissions    ┃
-┃    and limitations under the License.                                                            ┃
-┃                                                                                                  ┃
-┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┛
-                                                                                                  */
-package archimedes
-
-import anticipation.*
-import gossamer.*
-import vacuous.*
-
-// Token (leaf) elements: the elements whose content is character data rather
-// than child elements. `Mspace` and `Mglyph` carry no text at all (they are
-// controlled entirely by their attributes), so their `text` is `Unset`.
-
-sealed trait Token extends Mathml:
-  def contents: List[Mathml] = Nil
-
-case class Mi(value: Text, attributes: List[(Text, Text)] = Nil) extends Token:
-  def label: Text = t"mi"
-  def text: Optional[Text] = value
-
-case class Mn(value: Text, attributes: List[(Text, Text)] = Nil) extends Token:
-  def label: Text = t"mn"
-  def text: Optional[Text] = value
-
-case class Mo(value: Text, attributes: List[(Text, Text)] = Nil) extends Token:
-  def label: Text = t"mo"
-  def text: Optional[Text] = value
-
-case class Mtext(value: Text, attributes: List[(Text, Text)] = Nil) extends Token:
-  def label: Text = t"mtext"
-  def text: Optional[Text] = value
-
-case class Ms(value: Text, attributes: List[(Text, Text)] = Nil) extends Token:
-  def label: Text = t"ms"
-  def text: Optional[Text] = value
-
-case class Mspace(attributes: List[(Text, Text)] = Nil) extends Token:
-  def label: Text = t"mspace"
-  def text: Optional[Text] = Unset
-
-case class Mglyph(attributes: List[(Text, Text)] = Nil) extends Token:
-  def label: Text = t"mglyph"
-  def text: Optional[Text] = Unset
