+++ lib/archimedes/src/test/archimedes_test.scala
@@ -33,6 +33,7 @@
 package archimedes
 
 import soundness.*
+import Mathml.*
 
 import errorDiagnostics.stackTracesDiagnostics
 import strategies.throwUnsafely
